Vastly improve member search
Instead of searching the hierarchy view, a separate view of search results
is shown. This prevents confusing cases where a deep hierarchy is shown
and it is totally unclear which member was actually matched by the query.
In addition to this, search results are ordered by relevance.

For performance reasons, only a certain number of results are shown when
you type your query. Additional results must be enumerated by clicking the
"More results" button.